refactor(contracts): Remove address-names from exports
