+++ packages/contracts/deploy/008-Proxy__OVM_L1CrossDomainMessenger.deploy.ts
@@ -6,7 +6,7 @@ import {
   deployAndVerifyAndThen,
   getContractFromArtifact,
 } from '../src/hardhat-deploy-ethers'
-import { names } from '../src'
+import { names } from '../src/address-names'
 
 const deployFn: DeployFunction = async (hre) => {
   const Lib_AddressManager = await getContractFromArtifact(
