+++ packages/contracts/deploy/011-set-addresses.ts
@@ -6,9 +6,8 @@ import { DeployFunction } from 'hardhat-deploy/dist/types'
 import {
   getContractFromArtifact,
   isHardhatNode,
-  getContractFromArtifact,
 } from '../src/hardhat-deploy-ethers'
-import { names } from '../src'
+import { names } from '../src/address-names'
 
 const deployFn: DeployFunction = async (hre) => {
   const { deployer } = await hre.getNamedAccounts()
