+++ packages/contracts/src/index.ts
@@ -1,4 +1,3 @@
 export * from './contract-defs'
 export * from './predeploys'
 export * from './connect-contracts'
-export * from './address-names'
